fix for incorrect hotelling APU and shorepower emissions when running on Linux (#81)
The Go BaseRateCalculator reads input from flat files in the bundle. These flat files are always created with lowercase names. However, the go code was incorrectly trying to read the shorepoweremissionratefraction flat file using title case. This works on Windows, because file names are case insensitive on that platform. However, this fails on platforms where file names are case sensitive.